+++ app/database/migrations/2019_08_11_203612_remove_users_last_deleted_dominion_round_column.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RemoveUsersLastDeletedDominionRoundColumn extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up(): void
+    {
+        Schema::table('users', static function (Blueprint $table) {
+            $table->dropColumn('last_deleted_dominion_round');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down(): void
+    {
+        Schema::table('users', static function (Blueprint $table) {
+            $table->unsignedInteger('last_deleted_dominion_round')
+                ->nullable()
+                ->after('settings');
+        });
+    }
+}