+++ app/routes/web.php
@@ -138,9 +138,9 @@ $router->group(['middleware' => 'auth'], static function (Router $router) {
 
             // Op Center
             $router->get('op-center')->uses('Dominion\OpCenterController@getIndex')->name('op-center');
+            $router->get('op-center/clairvoyance/{realmNumber}')->uses('Dominion\OpCenterController@getClairvoyance')->name('op-center.clairvoyance');
             $router->get('op-center/{dominion}')->uses('Dominion\OpCenterController@getDominion')->name('op-center.show');
             $router->get('op-center/{dominion}/{type}')->uses('Dominion\OpCenterController@getDominionArchive')->name('op-center.archive');
-            $router->get('op-center/clairvoyance/{realmNumber}')->uses('Dominion\OpCenterController@getClairvoyance')->name('op-center.clairvoyance');
 
             // Government
             $router->get('government')->uses('Dominion\GovernmentController@getIndex')->name('government');
@@ -162,7 +162,6 @@ $router->group(['middleware' => 'auth'], static function (Router $router) {
             // Misc
             $router->post('misc/clear-notifications')->uses('Dominion\MiscController@postClearNotifications')->name('misc.clear-notifications');
             $router->post('misc/close-pack')->uses('Dominion\MiscController@postClosePack')->name('misc.close-pack');
-            $router->post('misc/delete-dominion')->uses('Dominion\MiscController@postDeleteDominion')->name('misc.delete-dominion');
 
             // Debug
             // todo: remove me later