+++ src/Calculators/Dominion/LandCalculator.php
@@ -131,28 +131,8 @@ class LandCalculator
         arsort($landPerType);
 
         $landLeftToLose = $totalLandToLose;
+//        $totalLandLost = 0;
         $landLostByLandType = [];
-
-        foreach ($barrenLandByLandType as $landType => $barrenLandForLandType) {
-            if ($landLeftToLose === 0) {
-                break;
-            }
-
-            $totalLandTypeLoss = $barrenLandForLandType;
-
-            if ($totalLandTypeLoss > $landLeftToLose) {
-                $totalLandTypeLoss = $landLeftToLose;
-            }
-
-            $landLostByLandType[$landType] = [
-                'landLost' => $totalLandTypeLoss,
-                'barrenLandLost' => $totalLandTypeLoss,
-                'buildingsToDestroy' => 0
-            ];
-
-            $landLeftToLose -= $totalLandTypeLoss;
-            $landPerType[$landType] -= $totalLandTypeLoss;
-        }
 
         foreach ($landPerType as $landType => $totalLandForType) {
             if ($landLeftToLose === 0) {
@@ -171,16 +151,21 @@ class LandCalculator
                 $totalLandTypeLoss = $landLeftToLose;
             }
 
-            if (!array_key_exists($landType, $landLostByLandType)) {
-                $landLostByLandType[$landType] = [
-                    'landLost' => 0,
-                    'barrenLandLost' => 0,
-                    'buildingsToDestroy' => 0
-                ];
+//            $totalLandLost += $totalLandTypeLoss;
+            $barrenLandForLandType = $barrenLandByLandType[$landType];
+
+            if ($barrenLandForLandType <= $totalLandTypeLoss) {
+                $barrenLandLostForLandType = $barrenLandForLandType;
+            } else {
+                $barrenLandLostForLandType = $totalLandTypeLoss;
             }
 
-            $landLostByLandType[$landType]['landLost'] += $totalLandTypeLoss;
-            $landLostByLandType[$landType]['buildingsToDestroy'] += $totalLandTypeLoss;
+            $buildingsToDestroy = $totalLandTypeLoss - $barrenLandLostForLandType;
+            $landLostByLandType[$landType] = [
+                'landLost' => $totalLandTypeLoss,
+                'barrenLandLost' => $barrenLandLostForLandType,
+                'buildingsToDestroy' => $buildingsToDestroy
+            ];
 
             $landLeftToLose -= $totalLandTypeLoss;
         }