+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 
 ## [Unreleased]
 
+## [0.7.0-3] - 2019-08-11
+### Changed
+- A bunch of empty realms now get created for each new round, to prevent people landing together when not packing
+- Land lost from being invaded is now again proportional to land types, including constructed/constructed buildings
+- Increased spy/wiz success rate
+- Race pages on the Scribes now show the race's home land type
+
+### Fixed
+- Units in training now count towards max military population
+
 ## [0.7.0-2] - 2019-08-09
 ### Added
 - Added link to scribes in the top navigation bar
@@ -365,7 +375,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Added
 - This CHANGELOG file.
 
-[Unreleased]: https://github.com/WaveHack/OpenDominion/compare/0.7.0-2...HEAD
+[Unreleased]: https://github.com/WaveHack/OpenDominion/compare/0.7.0-3...HEAD
+[0.7.0-3]: https://github.com/WaveHack/OpenDominion/compare/0.7.0-2...0.7.0-3
 [0.7.0-2]: https://github.com/WaveHack/OpenDominion/compare/0.7.0-1...0.7.0-2
 [0.7.0-1]: https://github.com/WaveHack/OpenDominion/compare/0.7.0...0.7.0-1
 [0.7.0]: https://github.com/WaveHack/OpenDominion/compare/0.6.2-9...0.7.0